feature(report): converts the mermaid plugin to a regular reporter
