+++ test/report/mermaid/__mocks__/contains-strings-to-be-escaped.json
