+++ test/report/mermaid/__mocks__/dependency-cruiser-2019-01-14.mmd
