+++ test/report/mermaid/__mocks__/do-not-follow-deps.mmd
